// src/models/Direccion.ts
export interface Direccion {
    iddireccion?: string;
    domisoci: string;
    provincia: string;
    canton: string;
    parroquia: string;
    sector: string;
    calleprincipal: string;
    numeracion: string;
    callesecundaria: string;
    descripcion: string;
    estado: string;
    created_at?: Date;
}